refactor: 環境変数由来の URL 定数を constants/urls.ts に集約する
lp/src/lib/web-url.ts と web/src/lib/lp-url.ts を constants/urls.ts に
移動し、LP 側は MiniChat の API_BASE も API_URL として同ファイルに
まとめる。今後の規約・行政向けページなど外部 URL 定数が増えたときの
見通しを良くする。
