+++ lp/src/constants/urls.ts
@@ -1,1 +1,2 @@
+export const API_URL = import.meta.env.PUBLIC_API_URL || "";
 export const WEB_URL = import.meta.env.PUBLIC_WEB_URL || "/";
